# backend/openapi.yml
openapi: "3.0.2"
info:
  title: Rest API Docs
  version: "1.0"
servers:
  - url: http://localhost:4444/
  - url: https://ciaro360-server.vercel.app/
components:
  schemas:
    ResponseWithMsg:
      type: object
      properties:
        status:
          type: boolean
        msg:
          type: string
...
    ResponseWithMsgError:
      type: object
      properties:
        status:
          type: boolean
          example: false
        msg:
          type: string
...
    Incomming_Tags:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: "6487f67cc0f0ec0734feaf64"
          name:
            type: string
            example: "DEMO2"
...
    Outgoing_Tags:
      type: object
      properties:
        listed:
          type: array
          items:
            type: string
          example: ["64410f9eb6f82aada0625a94", "64410f9eb6f82aada0625a94"]
        unlisted:
          type: array
          items:
            type: string
          example: ["Hr", "It"]
...
    getAllScopes:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        projects:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
        __v:
          type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /tag:
    get:
      tags:
        - "Tags Routes"
      description: "Get all tags"
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: count
          schema:
            type: integer
          description: Number of tags per page
        - in: query
          name: tagName
          schema:
            type: string
          description: Filter tags by name
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  tags:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incomming_Tags"
        "404":
          description: Tags not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized
      security:
        - bearerAuth: []
  /tag/{id}:
    get:
      tags:
        - "Tags Routes"
      description: "Get policies associated with a tag"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the tag
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: count
          schema:
            type: integer
          description: Number of policies per page
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  policies:
                    type: array
                    items:
                      type: string
                      example: "646085c40a5021e6d0805eae"
        "404":
          description: Tag not found or policies not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        401:
          description: Unauthorized

      security:
        - bearerAuth: []

    delete:
      tags:
        - "Tags Routes"
      description: "Delete a tag"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the tag
          schema:
            type: string
      responses:
        "200":
          description: Tag deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Tag not found or not deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        401:
          description: Unauthorized

      security:
        - bearerAuth: []
  /tag/{id}/{name}:
    put:
      tags:
        - "Tags Routes"
      description: "Update the name of a tag"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the tag
          schema:
            type: string
        - in: path
          name: name
          required: true
          description: New name for the tag
          schema:
            type: string
      responses:
        "200":
          description: Tag name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        "404":
          description: Tag not found or name not updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized
      security:
        - bearerAuth: []
  /tag/{id}/policies:
    delete:
      tags:
        - "Tags Routes"
      description: "Remove a tag from multiple policies"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the tag
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policies:
                  type: array
                  items:
                    type: string
                  description: List of policy IDs
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Tag not found or policies not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        401:
          description: Unauthorized
      security:
        - bearerAuth: []

  # Onboarding Routes
...
  /policy/{id}/tags:
    post:
      tags:
        - "Policy Routes"
      description: "Add tags to a policy"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the policy to add tags to
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                tags:
                  $ref: "#/components/schemas/Outgoing_Tags"
      responses:
        "200":
          description: Tags added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Adding tags to policy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        "401":
          description: Unauthorized
      security:
        - bearerAuth: []

    delete:
      tags:
        - "Policy Routes"
      summary: "Delete tags from a policy"
      description: "This endpoint deletes specific tags from a policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                  description: "List of tags to delete from the policy."
      responses:
        '200':
          description: "Tags deleted successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
...
  /risks/tags/{riskId}:
    get:
      summary: Retrieve tags for a risk
      tags:
        - Risk Routes
      parameters:
        - in: path
          name: riskId
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: true
          schema:
            type: integer
        - in: query
          name: count
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    default: true
                  tags:
                    type: array
                    items:
                      type: string
                    example: ["64981e9189aa814d71194e45","64981e9189aa814d71194e44"]
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      security:
        - bearerAuth: []
  
    post:
      summary: Add tags to a risk
      tags:
        - Risk Routes
      parameters:
        - in: path
          name: riskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Outgoing_Tags'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
      security:
        - bearerAuth: []
 
    delete:
      summary: Remove tags from a risk
      tags:
        - Risk Routes
      parameters:
        - in: path
          name: riskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags: 
                  type: array
                  items:
                    type: string
                  example: ["6496de9a100585e48f77c6dd","6496de9a100585e48f77c6de"]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
      security:
        - bearerAuth: []
...
  /scopes:
    post:
      summary: Create a new scope
      tags:
        - Scopes
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the scope.
                description:
                  type: string
                  description: The description of the scope.
              required:
                - name
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
        
    get:
      summary: Retrieve scopes
      tags:
        - Scopes
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            description: The page number.
        - in: query
          name: count
          schema:
            type: integer
            description: The number of items per page.
        - in: query
          name: scopeName
          schema:
            type: string
            description: Filter scopes by name.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAllScopes'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
  /scopes/{id}:
    put:
      summary: Update a scope
      tags:
        - Scopes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the scope to update.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The new name of the scope.
                description:
                  type: string
                  description: The new description of the scope.
              required:
                - name
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
    delete:
      summary: Delete a scope
      tags:
        - Scopes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the scope to delete.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
  /scopes/{id}/projects:
    post:
      summary: Link a scope to projects
      tags:
        - Scopes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the scope to link to projects.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
                  description: An array of project IDs to link to the scope.
              required:
                - projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
    delete:
      summary: Unlink a scope from projects
      tags:
        - Scopes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the scope to unlink from projects.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
                  description: An array of project IDs to unlink from the scope.
              required:
                - projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsg'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError' 
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []

  # EVIDENCE
...